SubscriptionsPage http implemented. Subscription Edit/Details Page finished. Logic changes on State Management. Minor model changes. Minor bug & design fixes.

+++ lib/content/providers/subscriptions.dart
@@ -2,7 +2,6 @@ import 'dart:math';
 
 import 'package:asset_flutter/common/models/response.dart';
 import 'package:asset_flutter/content/models/requests/subscription.dart';
-import 'package:asset_flutter/content/models/responses/subscription.dart';
 import 'package:asset_flutter/content/providers/subscription.dart';
 import 'package:asset_flutter/static/routes.dart';
 import 'package:asset_flutter/static/token.dart';
@@ -12,13 +11,7 @@ import 'dart:convert';
 import 'package:http/http.dart' as http;
 
 class SubscriptionsProvider with ChangeNotifier {
-  final List<Subscription> _items = [
-    Subscription('1', "Netflix 4K Family for Friends and Me", "Netflix Family Plan", DateTime.now(), BillCycle(month: 3), 40.5, 'TL', "netflix.com", 0xFFE53935),
-    Subscription('2', "Spotify", null, DateTime.now().subtract(const Duration(days: 5)), BillCycle(month: 1), 27.5, 'TL', "spotify.com", 0xFF4CAF50),
-    Subscription('3', "Playstation Plus", "Playstation Plus and this is an example of long text, lets see hot it'll behave.", DateTime.now().subtract(const Duration(days: 15)), BillCycle(year: 1), 165.2, 'TL', "playstation.com", 0xFF1976D2),
-    Subscription('4', "Jefit", null, DateTime.now().subtract(const Duration(days: 2)), BillCycle(day: 7), 10.9, 'USD', "jefit.com", 0xFF03A9F4),
-    Subscription('5', "WoW", null, DateTime.now().subtract(const Duration(days: 147)), BillCycle(year: 1), 60.2, 'USD', "worldofwarcraft.com", 0xFF4CAF50),
-  ];
+  final List<Subscription> _items = [];
 
   List<Subscription> get items {
     return [..._items];
@@ -26,6 +19,29 @@ class SubscriptionsProvider with ChangeNotifier {
 
   Subscription findById(String id) {
     return _items.firstWhere((element) => element.id == id);
+  }
+
+  Future<BaseListResponse<Subscription>> getSubscriptions({
+    String sort = "name", //name currency price
+    int type = 1
+  }) async {
+    _items.clear();
+    try {
+      final response = await http.get(
+        Uri.parse(
+          APIRoutes().subscriptionRoutes.subscriptionsByUserID + "?sort=$sort&type=$type"
+        ),
+        headers: UserToken().getBearerToken()
+      );
+
+      var baseListResponse = response.getBaseListResponse<Subscription>();
+      _items.addAll(baseListResponse.data);
+      notifyListeners();
+
+      return baseListResponse;
+    } catch (error) {
+      return BaseListResponse(error: error.toString());
+    }
   }
 
   Future<BaseAPIResponse> addSubscription(SubscriptionCreate subsCreate) async {
@@ -40,7 +56,7 @@ class SubscriptionsProvider with ChangeNotifier {
         _items.add(Subscription(
           Random().nextInt(9999).toString(), subsCreate.name, subsCreate.description, 
           DateTime.now(), subsCreate.billCycle, subsCreate.price, 
-          subsCreate.currency, subsCreate.image, subsCreate.color)
+          subsCreate.currency, subsCreate.image, subsCreate.color.toString())
         );
         notifyListeners();
       }
@@ -51,9 +67,25 @@ class SubscriptionsProvider with ChangeNotifier {
     }
   }
 
-  void deleteSubscription(String id) {
-    final deleteItem = findById(id);
-    _items.remove(deleteItem);
-    notifyListeners();
+  Future<BaseAPIResponse> deleteSubscription(String id) async {
+    try {
+      final response = await http.delete(
+        Uri.parse(APIRoutes().subscriptionRoutes.deleteSubscriptionBySubscriptionID),
+        body: json.encode({
+          "id": id
+        }),
+        headers: UserToken().getBearerToken()
+      );
+
+      if (response.getBaseResponse().error == null) {
+        final deleteItem = findById(id);
+        _items.remove(deleteItem);
+        notifyListeners();   
+      }
+
+      return response.getBaseResponse();
+    } catch (error) {
+      return BaseAPIResponse(error.toString());
+    }
   }
 }
